+++ packages/ui/package.json
@@ -31,6 +31,7 @@
 		"@radix-ui/react-navigation-menu": "^1.1.3",
 		"@radix-ui/react-popover": "^1.0.6",
 		"@radix-ui/react-slider": "^1.1.2",
+		"@radix-ui/react-slot": "^1.0.2",
 		"@tanstack/react-query": "^4.29.23",
 		"@tremor/react": "^3.4.1",
 		"@trpc/client": "^10.34.0",
@@ -39,6 +40,7 @@
 		"@trpc/server": "^10.34.0",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^1.2.1",
+		"cmdk": "^0.2.0",
 		"cva": "npm:class-variance-authority@^0.6.1",
 		"discord-markdown": "^2.5.1",
 		"highlight.js": "^11.8.0",
