+++ packages/ui/src/components/dashboard/domains.tsx
@@ -12,13 +12,15 @@ import {
 import { LuAlertCircle, LuXCircle, LuCheckCircle2 } from 'react-icons/lu';
 import { toast } from 'react-toastify';
 import { trpc } from '~ui/utils/trpc';
-import { Button, Input, LoadingSpinner } from '../primitives';
+import { LoadingSpinner } from '../primitives';
 import { useTierAccess } from '../primitives/tier-access-only';
 import { useDashboardContext } from './dashboard-context';
 import {
 	DomainVerificationStatusProps,
 	VercelDomainVerificationResponse,
 } from '@answeroverflow/api';
+import {Input} from "~ui/components/primitives/ui/input";
+import {Button} from "~ui/components/primitives/ui/button";
 
 function useDomainStatus({ domain }: { domain?: string }) {
 	const { data, isLoading, isFetching } =
