+++ packages/ui/src/components/pages/home/DemoMessage.tsx
@@ -4,10 +4,11 @@ import discordMarkdown from 'discord-markdown';
 import type { APIMessageWithDiscordAccount } from '@answeroverflow/api';
 import {
 	DiscordIcon,
-	Avatar,
-	AvatarImage,
-	AvatarFallback,
+
 } from '~ui/components/primitives';
+import {Avatar,
+  AvatarImage,
+  AvatarFallback,} from '~ui/components/primitives/ui/avatar';
 import Link from 'next/link';
 import type { ChannelPublicWithFlags } from '@answeroverflow/api';
 import { useIsUserInServer } from '~ui/utils/hooks';
