--- packages/ui/src/components/primitives/base/Button.tsx
@@ -1,66 +1,0 @@
-import * as React from 'react';
-import { type VariantProps, cva } from 'cva';
-
-import { cn } from '~ui/utils/styling';
-
-const buttonVariants = cva(
-	'active:scale-95 inline-flex items-center justify-center rounded-md text-sm font-medium transition-colors focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 dark:hover:bg-slate-800 dark:hover:text-slate-100 disabled:opacity-50 dark:focus:ring-slate-400 disabled:pointer-events-none dark:focus:ring-offset-slate-900',
-	{
-		variants: {
-			variant: {
-				default:
-					'bg-slate-900 text-white hover:bg-slate-700 dark:bg-slate-100 dark:hover:bg-slate-200 dark:text-slate-900 dark:hover:text-slate-900 ',
-				destructive:
-					'bg-red-600 text-white hover:bg-red-700 dark:hover:bg-red-700',
-				outline:
-					'bg-transparent border dark:border-slate-200 hover:bg-slate-200 border-slate-700 dark:text-slate-100 text-slate-900 dark:hover:bg-slate-800 dark:hover:text-slate-100',
-				subtle:
-					'bg-slate-100 text-slate-900 hover:bg-slate-200 dark:bg-slate-700 dark:text-slate-100',
-				ghost:
-					'bg-transparent hover:bg-slate-200 dark:hover:bg-slate-800 dark:text-slate-100 dark:hover:text-slate-100 data-[state=open]:bg-transparent dark:data-[state=open]:bg-transparent',
-			},
-			size: {
-				default: 'h-10 py-2 px-4',
-				sm: 'h-9 px-2 rounded-md',
-				lg: 'h-11 px-8 rounded-md',
-			},
-			selected: {
-				true: '',
-				false: '',
-			},
-		},
-		compoundVariants: [
-			{
-				variant: 'ghost',
-				selected: true,
-				className: 'bg-slate-200 dark:bg-slate-800',
-			},
-		],
-		defaultVariants: {
-			variant: 'default',
-			size: 'default',
-			selected: false,
-		},
-	},
-);
-
-export interface ButtonProps
-	extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-		VariantProps<typeof buttonVariants> {
-	selected?: boolean;
-}
-
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
-	({ className, variant, selected, size, ...props }, ref) => {
-		return (
-			<button
-				className={cn(buttonVariants({ variant, size, selected, className }))}
-				ref={ref}
-				{...props}
-			/>
-		);
-	},
-);
-Button.displayName = 'Button';
-
-export { Button, buttonVariants };
