--- packages/ui/src/components/primitives/base/Label.tsx
@@ -1,22 +1,0 @@
-'use client';
-
-import * as React from 'react';
-import * as LabelPrimitive from '@radix-ui/react-label';
-import { cn } from '~ui/utils/styling';
-
-const Label = React.forwardRef<
-	React.ElementRef<typeof LabelPrimitive.Root>,
-	React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>
->(({ className, ...props }, ref) => (
-	<LabelPrimitive.Root
-		ref={ref}
-		className={cn(
-			'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
-			className,
-		)}
-		{...props}
-	/>
-));
-Label.displayName = LabelPrimitive.Root.displayName;
-
-export { Label };
