+++ packages/ui/src/components/primitives/base/LinkButton.tsx
@@ -2,7 +2,8 @@ import type { VariantProps } from 'cva';
 import Link from 'next/link';
 import type { SetRequired } from 'type-fest';
 import { cn } from '~ui/utils/styling';
-import { buttonVariants } from './Button';
+import {buttonVariants} from "~ui/components/primitives/ui/button";
+
 
 export interface LinkButtonProps
 	extends SetRequired<
