+++ packages/ui/src/components/primitives/base/index.ts
@@ -1,18 +1,6 @@
-export * from './Avatar';
-export * from './Button';
-export * from './Dropdown';
 export * from './Heading';
 export * from './Icons';
-export * from './Input';
-export * from './Label';
 export * from './LinkButton';
-export * from './Menubar';
-export * from './NavigationMenu';
 export * from './Paragraph';
-export * from './Popover';
 export * from './Link';
 export * from './LoadingSpinner';
-export * from './Skeleton';
-export * from './Accordion';
-export * from './Dialog';
-export * from './AlertDialog';
