+++ packages/ui/src/components/primitives/ui/skeleton.tsx
@@ -1,0 +1,15 @@
+import { cn } from "~ui/utils/utils"
+
+function Skeleton({
+  className,
+  ...props
+}: React.HTMLAttributes<HTMLDivElement>) {
+  return (
+    <div
+      className={cn("animate-pulse rounded-md bg-muted", className)}
+      {...props}
+    />
+  )
+}
+
+export { Skeleton }
